+++ lib/matestack/ui/vue_js/registry.rb
@@ -6,6 +6,7 @@ module Matestack
         Matestack::Ui::Component.register(
           toggle: Matestack::Ui::VueJs::Components::Toggle,
           onclick: Matestack::Ui::VueJs::Components::Onclick,
+          transition: Matestack::Ui::VueJs::Components::Transition,
         )
 
       end