+++ spec/dummy/app/controllers/demo_controller.rb
@@ -7,4 +7,8 @@ class DemoController < ActionController::Base
     render Demo::FirstPage
   end
 
+  def second
+    render Demo::SecondPage
+  end
+
 end