+++ spec/dummy/app/matestack/demo/first_page.rb
@@ -4,6 +4,9 @@ class Demo::FirstPage < Matestack::Ui::Page
     # Demo::Components::Header.()
     header slots: { first: a_slot(1), second: a_slot(3) }
     h1 'First page with new logic!'
+    transition path: second_path do
+      button 'Second Page'
+    end
     abbr title: 'test'
 
     rails_render partial: '/some_partial', locals: { foo: 1 }