+++ spec/dummy/app/matestack/demo/second_page.rb
@@ -1,0 +1,10 @@
+class Demo::SecondPage < Matestack::Ui::Page
+
+  def response
+    h1 'Second Page'
+    transition path: root_path do
+      button 'First Page'
+    end
+  end
+
+end